move smartstack stuff into smartstack_tools
